# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ellipsonic OpsHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/recent-activity:
    get:
      operationId: getDashboardRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/upcoming-deadlines:
    get:
      operationId: getDashboardUpcomingDeadlines
      tags: [dashboard]
      summary: Get upcoming report deadlines
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Upcoming deadlines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeadlineItem"
  /dashboard/chart-data:
    get:
      operationId: getDashboardChartData
      tags: [dashboard]
      summary: Get chart data for dashboard
      security: [{ bearerAuth: [] }]
      parameters:
        - name: weeks
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardChartData"

  # ───── DEPARTMENTS ─────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DashboardStats:
      type: object
      properties:
        total_employees:
          type: integer
        total_departments:
          type: integer
        pending_reports:
          type: integer
        approved_reports:
          type: integer
        late_reports:
          type: integer
        weekly_completion_pct:
          type: number
        reports_this_week:
          type: integer
        reports_submitted_today:
          type: integer
      required:
        [
          total_employees,
          total_departments,
          pending_reports,
          approved_reports,
          late_reports,
          weekly_completion_pct,
          reports_this_week,
          reports_submitted_today,
        ]
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        user_name:
          type: string
        user_photo:
          type: ["string", "null"]
        action:
          type: string
        entity_type:
          type: string
        entity_id:
          type: string
        description:
          type: string
        created_at:
          type: string
      required: [id, user_name, action, entity_type, entity_id, description, created_at]
    DeadlineItem:
      type: object
      properties:
        department_id:
          type: string
        department_name:
          type: string
        deadline:
          type: string
        days_remaining:
          type: integer
        submitted_count:
          type: integer
        total_count:
          type: integer
      required:
        [department_id, department_name, deadline, days_remaining, submitted_count, total_count]
    DashboardChartData:
      type: object
      properties:
        weekly_reports:
          type: array
          items:
            $ref: "#/components/schemas/TrendDataPoint"
        department_comparison:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentComparisonItem"
      required: [weekly_reports, department_comparison]
    TrendDataPoint:
      type: object
      properties:
        label:
          type: string
        submitted:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        late:
          type: integer
      required: [label, submitted, approved, rejected, late]
    DepartmentComparisonItem:
      type: object
      properties:
        department_id:
          type: string
        department_name:
          type: string
        submitted:
          type: integer
        approved:
          type: integer
        completion_pct:
          type: number
        total_employees:
          type: integer
      required:
        [department_id, department_name, submitted, approved, completion_pct, total_employees]

    # ── Departments ──
...